Separated rendering from rasterization
This refactors the layout of the crates to introduce a "rasterized"
frontend. The Rasterized frontend takes a Renderer, which is a trait
introduced in Core. Kludgine implements Renderer, and the Rasterized
frontend implements cross-platforom Transmogrifiers using the Renderer
abstraction.

This Renderer trait can also be used to implement a Canvas API
eventually, which also means that it could be possible to use rasterized
widgets within a Canvas inside of any other frontend.

This commit also removes all layout code. In talking with Daxpedda, we
decided it was best to separate the roles of laying out a UI and widget
handling. The layout code will eventually consume this widget code to
produce cross-platform layouts.