Implemented a WebSys frontend.
Also started experimenting with an xtask setup. In this design, it's
thought that a Box widget's transmogrifier would be responsible for
calling transmogrify on the contained widget. For WebSys, this returns
the created element, which would allow the parent widget to manage that
handle.

The only problem is that the current design offers no way to store
information for transmogrifier. I believe another associated type should
work nicely for this, but I'll deal with this challenge in another
commit.