Updating based on feedback session.
In browser, we now are using the convention of using the most specific type inside of the transmogrifier, and then downcasting into Element.

The Surround code has been updated to not use DImension anymore, which makes the implementation consistent between the platforms -- Centered after adding padding. Container should still support setting different alignments, but it's now consistent with the available features.