Initial eventing hooked up
This implementation attempts to avoid background tasks by using channels
and requiring that the frontend invokes "process_events" on Gooey when
it makes sense for that frontend.

There isn't any actual usage of this plumbing yet -- it was a chore to
get the generic type handling working so that this much would simply
compile. I then cleaned up the code a little more, and decided it's a
good checkpoint for this project.