Browser button clicking "works"
This is a bit ugly but it's coming together. I think for usability I
need to try to merge some parameters together.

This chunk of works allows the button to register a closure for onclick
which passes an event to the Widget. The widget can respond with a
command to the transmogrifier.

After I finished this, I realized I wanted to think a bit more about the
logistics here. This method makes it easy to register a Js closure that
emits an event to the widget. But, maybe there should be an easy-to-use
mechanism to receive a inenral message to your transmogrifier. You can
do it by hand the same awy that WidetClosure is working, but it seems
like it could be a common use case.