Promoting frontend.
My fear of putting frontend on the root transmogrifier trait was unfounded -- it's the Widget trait that shouldn't know about the frontend.

This argument list is a bit gnarly, but I think it's the final list. The name of Channels should change, but each argument has a role and a reason it shouldn't be stored together. An argument could be made to create a borrowed type that further borrows these arguments so that they can be passed as a single argument -- I may still do that.