Counter example works on wasm.
This was a nightmare to implement. It's been two days of fighting the
architecture to allow communication between widgets. The underlying
issue is that the widget layer shouldn't know about the frontend, at
all. The problem is, the frontend is required for certain actions,
including sending an event into the runtime.

This led to a conundrum: how do you write cross platform code that uses
a frontend, but doesn't know anything about the frontend?

The answer is... a lot of Any and a lot of abstraction. This isn't by
any imagination a *good* implementation. The design that is working only
materialized this morning after more banging of my head against a wall.

Ultimately, the biggest win was moving Channels into WidgetState from
Transmogrifier state. There's a lot of cleanup that can still happen
because of this change, but I wanted to get this committed since it's in
a working state. It's much easier to review my progress when I'm working
from known states.