Restored example to having a Container.
Basic isn't really basic. It's just a testbed for now. We need a full
form setup to make a real example, and we're getting closer to that. I'm
confident implementing this sort of functionality in my idea of a form
will be more concise and easier to read.

This fixes the issues that were limiting the ability to keep the basic
example using a Container wrapping the Button. Until late in the design
of the eventing system, there wasn't a way to directly message another
widget if you knew its type and ID. There is now, and so another
redesign of widget creation had to be looked at.

Container can now be initialized directly with a widget and storage, or
it can be given a widget registration.

Despite these simplistic-looking beginnings, I think we have everything in place
necessary to implement this framework at the cross-platform level. I
think it's now a matter of refinement and cleanup, but the majority of
the remaining work will be in the frontends and widget code.