Refactored rendering and clipping.
Needed a place to note when each widget is drawn, so that I can record
the order and locations and IDs of each of those widgets. This
render-map is what will be needed to look up what widget should be sent
mouse events, for example.

In the process of doing this, I wanted to make each widget's render call
be rendered at 0,0,w,h, and the view port is already clipped and located
for your drawing calls.