Button clicks now work on rasterized frontends.
While hover state is tracked and events fire, there no support for doing
anything with those state changes. This just solves the need of handling
event routing on winit-based frontends. The flow for a direct-winit
consumer will be to convert the winit WindowEvent into the
gooey_rasterizer::WindowEvent through the built-in method. If it returns
the event back to you, it's not an event Gooey can handle. If you get a
Gooey WindowEvent, you pass that onto the Rasterizer and it will tell
you if it handled the event or not. There's an ownership issue that I
just realized as I type this, but it's not far off from supporting the
flow of checking if a winit mouse click was handle by gooey, and if not,
letting your game engine handle it instead.

For Klugine, it translates the winit event into its own internal event
types to send across a channel to the worker thread that handles the
window event loop. I couldn't come up with a design that allowed me to
move that handling out of Kludgine, so instead of using Winit -> Gooey
event conversion, the kludgine_main implementation handles it for you.
I'd like to move that somewhere else, in case you don't want to use
kludgine_main but are still using Kludgine.

In Kludgine, the State was represented by two parts: layers and the base
state. This implementation just puts all the state in one place for now,
as the plan of action for layers isn't well defined yet.