Implemented styles
This is a rough first pass. The system works, and this commit is to
serve as a checkpoint of a working state as well as offer a chance to
get feedback on the pros and cons of the current design.

The first problem was, how do you allow the user to specify style
information during widget creation? My approach was to add a "with"
method that accepted a StyleComponent and returns a StyledWidget struct.
Or, if you want to just accept the default style, you call
`with_default_style`. Honestly, I'm happy with this design, but
with_default_style is a bit verbose.

The second problem was, how do you allow the user to provide a global
stylesheet that's used regardless of frontend, but also allow individual
frontends to have their own customizable themes as well? The
kludgine_main functions don't offer a way to set this stylesheet, but
other than figuring out a way to expose it at the App level, this system
works. However, I think I might want to incur the overhead of a mutex
and allow the Rasterizer to update the shared stylesheet during
construction.

But, with those notes, the button in the rasterizer example now changes
colors when hovered.