More CSS support
* Applying styles directly to a widget now works.
* Class rules, like the newly introduced `gooey-root` rule, are now
applied by the WebSys transmogrifier inline.

The result of these changes is that both apps now share a common
background color, defined by the stylesheet. Still, the browser version
needs some more reflection.

Ultimately, if I want the browser to apply the CSS rules for class-only
rules (ones where no widget is specified), either it can only support
the basic CSS types and never be expanded, *or* the rules have to
applied as they are in this branch -- as id-driven rules on a per-widget
basis.

This implementation still isn't complete -- class only rules can't
currently restrict based on state. It's more of a proof of concept of
how this might have to work with CSS.