Implemented 'custom layout' for browser.
Still not happy with the name for this layout approach. Most people will
want a grid, but sometimes you'll want the equivalent of CSS's
Relative/Absolute approach. That's what this layout widget attempts to
provide: arbitrary list of widgets laid out using
left/top/right/bottom/width/height relative to the layout widget's
bounds.

The rasterizer layer is unimplemented in this commit.

The browser example HTML now dynamically loads the example based on the
query string, and the xtask command now supports naming which example to
build for the browser. This makes it much easier to switch between
examples, to the point that having a dropdown to switch examples
wouldn't be too crazy to add.