Enabled out-of-band messaging to widgets.
Context can now be cloned and sent to other threads/async tasks and used to
communicate back to Gooey. Currently only implemented for Rasterizer
since the demo app is only working on the rasterized version. Shouldn't
be hard to hook up on wasm.

The frontend is now where widgets should be marked as having messages
pending. The frontend's method can optionally take extra steps to wake
up the Gooey's event loop. For rasterizer, it now has an optional
callback that can be installed for such an occasion. The Kludgine App
implementation calls Kludgine's's new
RedrawRequester::set_needs_update().

The net effect is allowing external threads to send messages into
Gooey's ecosystem without triggering frames to be drawn.

For wasm, as long as we're single-threaded, we can simply proceed
to processing messages.