Imported the pliantdb-counter example.
As I started structuring an mdbook, I realized I would want to tie a
tutorial about how this would be built, step-by-step, by importing the
files into the mdbook.

The best way to do that seemed to be to import the example. Because
these examples are multi-crate monstrosities, I created a separate
folder for them to live in. It didn't seem prudent to try placing these
in `gooey/examples`. I imagine we'll have several examples showing
various combinations of usages, and this folder can be for the complex
ones that require more than one crate to implement.

Technically, this example could be one crate with multiple feature
flags, but I feel it's a requirement of this example to show how we
expect developers to build their applications, and it's going to be
suggesting separating these concerns into separate crates like we do in
the example.