Removed Widget Commands in favor of direct access.
As I was starting the book, I was wantin gto write a chapter dealing
with communication. That made me revisit how external widgets
communicated, and I realized that I might be able to do away with
messaging entirely in favor of direct access. There are limitations, due
to needing to ensure having access.

This does raise a potential issue of the lock type being used here.
Right now it's std::sync::Mutex, and the pliantdb-couner-example is
using a context to invoke `with_widget` from an async task. This means
that it's possible to cause the async tasks' thread to go to sleep while
waiting for the Gooey thread to execute some code, but since the Gooey
thread is its own thread, this technically is "ok" if not perfectly
desired. Regardless, the lock contention should be very infrequent, so
the situation should rarely arise.

The results of the updates to the examples speak for themselves, I think.