Removed alignment support for checkbox.
I'm not really sure what the alignment support truly meant. I'm having
trouble coming up with real-world use cases of a checkbox to the right
of its label, when not used within a larger form context. In that
situation, I woudln't expect the label to be attached to the actual
checkbox anyways.

Also switched the check to be a box, which should be easier to replicate
in css.