Custom styling of checkbox in the browser
"Checkbox should be easy," I thought. I regret my thoughts yesterday.
Getting the styling to work the same way for checkboxes and radios are
annoying. Hopefully this can be refactored some to share code when we
add radio buttons.

I standardized a method for dealing with the theme/state css rendering,
as the checkbox needed it. I added an `sr-only` class that is
automatically added by the websys frontend which hides an element
visually while ensuring it still shows up to accessibility.

The rest of the implementation is all just making a fake checkbox work.
Phew.