Refactored component content creation
This addresses a pain point I realized while working on the widget
explorer: it was too easy to forget to register a widget in both the
layout and the component. This commit formally introduces the builder
pattern for widget initialization.

The net effect is that widgets must implement the Content interface to
be used as the Content of a Component. This requires your builder to use
KeyedWidgetStorage instead of WidgetStorage directly. The only downside
of this change is the more involved type signatures when implementing
one of these container-style widgets. But, the payoff for the end-user
not needing to remember to register their widget in multiple locations
is worth it.