Added RelatedStorage
This new trait adds the last link allowing someone to implement a new
Component widget. I was trying to avoid a locking structure for the
widget storage within Component, but I couldn't avoid it. The component
calling functions that register/remove widgets will already have the
component locked, so a deadlock occurs if the contained widget turns
around and tries to update the component. By reducing the lock to the
widget hash map, it allows the code to function properly.

The net effect of this is that calling Layout::insert will now
automatically insert the widget into the containing component, if
present and if a key is specified.