Refactored colors
Colors are now mostly const fn operations internally, and only convert
to Palette as needed, which right now is never. However, we will be
using palette to do lightening/darkening eventually.

Since the role of palette was only to be a way to safely do color
operations, we don't need to prioritize storing the data using palette's
types. Due to palette not being able to support const fn yet, it seemed
prudent to switch the storage type so that we could introduce named color
constants that are true constants.