Labels now use Text internally
Additionally, text rendering now accepts a context style. Since text
itself can be styled, the context style is used to fill in the blanks
when a value can't be found in the span's style. Thus, a label's text
color setting will flow into the Text when rendering without needing to
alter the underlying spans.

I didn't update any examples with this functionality. It's a
work-in-progress to adding a markdown conversion function that will make
a better demo.