Added Border and Padding
This refactors Surround a little bit to allow it to be used across all
of these types. Moved button's built-in padding to the style sheet.
Lastly, for reasons necessary at the time, I refactored all
transmogrifier contexts to be passed as &mut. In the end, I didn't need
to call multiple methods with the same context, but this allows it.