Initial commit for input widgets
The initial goal is a very basic text field that supports
setting/getting the value, events when the value changes.

The browser is implemented, but completely untested.

The rasterizer is being slowly reconstituted from the old Kludgine code.
I've reached a point where I wanted to start working on a hit list,
although this is still very much a work in progress.