Switched to figures.
This probably should have been done on main... but... too late! I was
craving API improvements while working on this project, hence why the
project arose. This was the bare minimum to maintain functionality after
an initial API implementation. I expect to be able to simplify some of
the existing code already but over time with additional features as
well.