Implemented Builder for Input
Added a basic example, although it doesn't really do anything except let
you type in it. For now, didn't add an actual demo because it's ugly
with red. There's too many things I want to do, but so much is unrelated
to input specifically, which is why I wanted to finish up this PR even
though it's a big rough around the edges.