Implemented LocalizableError
Added an easy way to use `Display` for pre-existing errors that you may
not want to translate. For errors that are from foreign crates, I
included a macro that does the same implementation. This should make it
nearly painless to use Gooey without localization, while still making a
path for full localization support possible.