Added default/cancel support as Intents.
The browser support is a little hacky -- the hierarchy exists in the
dom, and we probably should reconstruct it to resolve the default/cancel
handlers. Instead, we use a "last constructed" approach, which means
that if you have multiple submit buttons on the same layer, it might be
what you intend or it might not.

There are better ways to handle it, but they're all much more
complicated. And, for now, I think it's ok to suggest not putting
multiple submit/cancel intents on the same layer.