Fixed unadorned label iterator
The previous implementation of this iterator was unwrapping the outer
option, leaving the inner value as None. This was in essence causing
Some(None) output from the parent iterator to always produce None in the
output.

This new implementation keeps the Some(None)s intact, allowing for the
previous workaround of flatten to be changed back to unwrap.

Refs #51