Fixed lock() api in non-managed code.
In the bonsaidb counter example, the callback that comes back from
BonsaiDb is executed outside of managed code. Without the block having a
managed code guard active, set_label will cause widget events to fire
immediately, which attempts to lock the widget. If it can't acquire the
lock, parking_lot kicks in.

In WASM, parking_lot isn't available, hence a crash that was happening
in the example. The widget guard now contains a managed code guard to
prevent this from happening.