Updating readme steps
I didn't read my own steps when executing them, and 'fixed' the
miniserve command. In the end, I think most users will be in the
repository root when running these commands. So, I updated that text
instead of changing the command back.

The xtask command generates the proper path to suggest launching
miniserve with, regardless of where you invoke it.