Partial Flex implementation
This is a bit of a doozie. The last commit allowed gooey-kludgine to
draw the root widget, but when I got to implementing the Flex widget, I
realized that there was no way to prove to the compiler that the generic
Renderer parameter matched the one that it was originally boxed for.

The reason I can't use associated types for this relationship ultimately
boils down to rust-lang/rust#100013. The clipped associated type
complicated things too much.

This morning I had the idea to force the Renderer trait into being
object safe by forcing implementations for Px and Lp rather than being
fully generic. The implementations can still specify a generic parameter
for the Unit type, but now the type can be boxed with exactly two
generic types supported: Px and Lp. I then gave up on the goal of having
a ClipGuard for the renderer type, and it finally was fully object safe.