Added widget measurement
This isn't quite the best API yet, but it proves that the functionality
works. Taffy is a bit of a pain to integrate due to the MeasureFunc
requiring Send + Sync + 'static, since it gets boxed. As such, this
delegated-through-a-thread-and-channels approach is the only way for
interactive measurements to access the local graphics context.

This makes me think that the original design of Gooey was on the right
track: skip flexbox and provide purpose-built widgets a la Flutter.