Replaced Taffy with custom implementation
After having gotten the Taffy code working, I wasn't very happy with the
fact I needed a thread to implement it due to the lifetime requirements
on the Graphics contexts. This meant that Gooey's raster code wouldn't
work on WASM without extra work.

After looking a bit at Flutter and SwiftUI, I realized that my original
inclination in the previous iteration of Gooey was still the best
approach: purpose-built layout classes rather than the super complicated
flex and grid system that the browser uses.

This commit adds a basic implementation of a flex container that
supports laying children out in either columns or rows. The web's
implementation still doesn't actually assign any CSS for the layout, so
the platforms behaviors are still not lined up, but now the raster's
implementation is more correct than the browser's.