Fixing button animations being "slow"
The issue was that my last set of changes were causing the animations to
restart, causing the animation to keep being extended to another 150ms.

I think the only way for this to work is to switch to an event
mechanism to notify widgets once they've been invalidated. This event
could include a parameter stating whether it was a direct invalidation
or an invalidation due to another widget in the hierarchy. Button
doesn't really care about the rest of the hierarchy, it only cares about
its own state, and the cache key was including too many changes.