Added WidgetLayout + Baseline
Refs #88

This set of changes starts working towards supporting baseline
alignment. The Grid widget currently does not honor vertical align in
any way, but it in theory propagates the first row's baseline
information -- which will need to be adjusted if Baseline vertical
alignment is used.

Next step is to actually add Baseline and try to make it work. The set
of changes thus far was just so massive and I finally resolved all the
errors from this refactoring.