Disclose now forces size to fit vertically
The new Label behavior of honoring the `Fill` constraint caused the
Disclose widget to misbehave when a Label was used as the label when the
Disclose was being asked to fill a given space.

Since the Disclose widget is always vertical, it can make the assumption
it should overrule the layout requirements and always fit the label
vertically.