Label now invalidates with more settings
From discord, the text size being set for a label wasn't always causing
the label to be laid out correctly. This ended up being caused by the
text size not being included in the label cache key. I've refactored
this to support all of the font settings, and in theory make it harder
to break in the future.