Added shaders-full example + batched render ops
This commit adds a new example showing an approach to handling
ortographic projection into the area a widget occupies. It also shows
how to handle opacity.

There was a bug in how opacity was being handled -- Cushy's opacity
wasn't being provided to the render operation.

Lastly, while working on this I was inspired to offer a way for render
operations to support batching. The main changes for this are in
Kludgine, but `RenderOperation::batch_prepared` is the new function that
supports this in Cushy.