Enforce one-execution-per-chain
One of the fallouts of executing callbacks in a deferred state is that
the locks no longer prevent cyclic updates. For well defined values,
this works perfectly. But for something like floating points, this can
cause not-quite-the-same values to continue to propagate until they
stabilize. The 7guis-temperature-converter, for example, would
sometimes replace the currently editing text with an imprecise float
result computed from a roundtrip temperature conversion.

This change adds tracking to these deferred callbacks that ensures for
each "root" invocation of a dynamic's callbacks, all affected change
callbacks are invoked a single time in any given execution chain, taking
care to support divergent chains causing an individual set of callbacks
to be invoked more than one time as long as their individiaul call
chains have not yet invoked the callbacks.