Add Tracked::ignoring_first_if
This also avoids setting the window's icon to None by using this
feature. Originally this was meant to be a workaround for a bug in
winit, but the bug seems to already be fixed in winit 0.30.8. Still,
this reduces some interthread communication when opening a window so
it's a good change!