Dialog buttons no longer always close the dialog
Dialog button callbacks now return a ShouldClose and are FnMut instead
of FnOnce. This allows nested dialogs to be invoked through the buttons
on a dialog in addition to the content of the dialog itself.

While doing this, I realized the documentation for
ModalHandle::build_dialog claimed it was used to create nested dialogs,
but it wasn't actually implemented that way. In trying to make the
modal.rs expanded example work, I wanted both behaviors -- updating the
current handle contents with a new dialog and creating a nested dialog,
so there is now a `ModalHandle::build_nested_dialog` function in
addition to the existing function.