Initial Channel implementation
This new type isn't integrated in with any of the existing types yet.
This is a step towards what is needed to implement the command pattern,
which I think is the best approach for #192.

I think we might need a Broadcast channel type as well to allow a
channel type where multiple callbacks can be associated rather than a
single, but it will require a Clone bound or extra allocations and the
current implementation doesn't need either.